third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-reconnect-during.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let device, fake_peripheral, service, fake_service;
  return getHealthThermometerDeviceWithServicesDiscovered({
      filters: [{services: [health_thermometer.name]}],
  })
    .then(_ => ({device, fake_peripheral} = _))
    .then(() => fake_peripheral.addFakeService({uuid: health_thermometer.name}))
    .then(s => fake_service = s)
    .then(() => device.gatt.getPrimaryService(health_thermometer.name))
    .then(s => service = s)
    .then(() => fake_service.addFakeCharacteristic({
      uuid: measurement_interval.name,
      properties: ['read', 'write', 'indicate'],
    }))
    .then(() => service.getCharacteristic(measurement_interval.name))
    .then(characteristic => assert_class_string(
        characteristic, 'BluetoothRemoteGATTCharacteristic'))
    .then(() => {
      let promise = assert_promise_rejects_with_message(service.getCharacteristics(), new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));
      // connect() guarantees on OS-level connection, but disconnect()
      // only disconnects the current instance.
      // getHealthThermometerDeviceWithServicesDiscovered holds another
      // connection, so disconnect() and connect() are certain to reconnect().
      // However, disconnect() will invalidate the service object so the
      // subsequent calls made to it will fail, even after reconnecting.
      device.gatt.disconnect();
      return device.gatt.connect().then(() => promise);
    });
}, 'disconnect() and connect() called during getCharacteristics. Reject with ' +
   'NetworkError.');

</script>
